refactor : repackaging table
talbe 관련 클래스 리패키징
