+++ src/main/java/kitchenpos/table/dao/OrderTableDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.table.dao;
 
 import kitchenpos.domain.OrderTable;
